fix: use host semantic-release configuration
Closes #11

Previously, next-release-version deleted the host's semantic-release
configuration before invoking semantic-release because this action
doesn't install the host's node_modules and semantic-release throws a
runtime error when the `plugins` configuration references a plugin that
can't be `require`d.

This was too aggressive an approach, because deleting the host's entire
semantic-release configuration removed critical configuration like
`branches`. When a repository uses main instead of master and the
default semantic-release `branches` configuration is used, the lack of a
master branch causes an `ERELEASEBRANCHES` error.

This commit keeps the host semantic-release configuration but overrides
the `plugins` setting using the `--plugins` CLI option[^1].

[^1]: https://semantic-release.gitbook.io/semantic-release/usage/configuration#plugins