fix: override plugins for each release step
Relevant documentation:
https://semantic-release.gitbook.io/semantic-release/usage/plugins#plugins

In #11, @RampedIndent found that `semantic-release --dry-run` tries
`require`-ing plugins for all steps, even the
prepare/publish/success/fail steps that are not executed in `--dry-run`.

This commit uses the CLI options to override the configured plugins
for all steps, with the exception of analyze-commits.

That means this GitHub Action does not support semantic-release with
a non-stock analyze-commits configuration. Punt that issue until it's
observed in the wild.